Integrate join template, find auth template to auth template
