+++ src/components/templates/AuthTemplate/AuthTemplate.styled.ts
