--- src/components/templates/JoinTemplate/JoinTemplate.styled.ts
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledJoinTemplate = styled.div`
-  padding: 2.5rem 0;
-  position: relative;
-`;
-
-export const StyledHeaderSection = styled.section`
-  margin-bottom: 2.5rem;
-`;
-
-export const StyledJoinFormSection = styled.section`
-  @media (max-width: 769px) {
-    margin: 0 2rem;
-  }
-`;
